🔧 rename (app/autocomplete/bash_autocomplete, app/autocomple...)
