+++ app/cmd/alias/delete.go
@@ -4,8 +4,8 @@ import (
 	"sync"
 
 	"github.com/HidemaruOwO/nuts/log"
-	"github.com/HidemaruOwO/pummit/src/config"
-	"github.com/HidemaruOwO/pummit/src/lib"
+	"github.com/HidemaruOwO/pummit/app/config"
+	"github.com/HidemaruOwO/pummit/app/lib"
 )
 
 func AliasDeleteCmd(targets []string) {
